Created a read_fixture(...) function in the tests package __init__.py
This type of functionality, reading static response content from the
fixtures directory is done in many unit tests in some form or another.
I refactored test_pei.py to use it and intend on using it with
test_nspower.py for NSPowerClient.

We can gradually refactor other classes to make use of the centralized
function as well.
